Better cast type for pointers
- also fix indentation

+++ src/core/StelViewportEffect.cpp
@@ -338,11 +338,11 @@ void StelViewportDistorterFisheyeToSphericMirror::setupCurrentVAO() const
 	auto& gl = *QOpenGLContext::currentContext()->functions();
 	verticesVBO->bind();
 	gl.glVertexAttribPointer(VERTEX_ATTRIB_INDEX, SPHERIC_MIRROR_COORDS_PER_VERTEX, GL_FLOAT,
-				 false, 0, reinterpret_cast<const void*>(qint64(vboVertexDataOffset)));
+	                         false, 0, reinterpret_cast<const void*>(uintptr_t(vboVertexDataOffset)));
 	gl.glVertexAttribPointer(COLOR_ATTRIB_INDEX, SPHERIC_MIRROR_COORDS_PER_COLOR, GL_FLOAT,
-				 false, 0, reinterpret_cast<const void*>(qint64(vboColorDataOffset)));
+	                         false, 0, reinterpret_cast<const void*>(uintptr_t(vboColorDataOffset)));
 	gl.glVertexAttribPointer(TEXCOORD_ATTRIB_INDEX, SPHERIC_MIRROR_COORDS_PER_TEXCOORD, GL_FLOAT,
-				 false, 0, reinterpret_cast<const void*>(qint64(vboTexCoordDataOffset)));
+	                         false, 0, reinterpret_cast<const void*>(uintptr_t(vboTexCoordDataOffset)));
 	verticesVBO->release();
 	gl.glEnableVertexAttribArray(VERTEX_ATTRIB_INDEX);
 	gl.glEnableVertexAttribArray(COLOR_ATTRIB_INDEX);
